Add persistent base and create models dir
